+++ src/components/Popover/index.js
@@ -1,0 +1,3 @@
+import Popover from './Popover.react';
+
+export default Popover;