+++ src/index.js
@@ -11,6 +11,7 @@ import FlatButton from './components/FlatButton';
 import FontIcon from './components/FontIcon';
 import IconButton from './components/IconButton';
 import Paper from './components/Paper';
+import Popover from './components/Popover';
 import Snackbar from './components/Snackbar';
 import Subheader from './components/Subheader';
 import RadioButtonGroup from './components/RadioButtonGroup';
@@ -34,6 +35,7 @@ export {
   FontIcon,
   IconButton,
   Paper,
+  Popover,
   Snackbar,
   Subheader,
   RadioButtonGroup,