+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## 2.16.1 - 2018-09-05
+### Fixes
+- Fixes https://github.com/StratoDem/sd-material-ui/issues/79 where `Checkbox` callback was firing twice to the server on a click event
+
+## 2.16.0 - 2018-08-30
+### Added
+- Adds `Popover` as an uncontrolled component which can render other components as children
+
 ## 2.15.0 - 2018-06-13
 ### Added
 - Adds `searchEndpointAPI` prop to `AutoComplete` to allow for external searching https://github.com/StratoDem/sd-material-ui/issues/76